Refactor Markdown Exporter
